evaluation/requirements.txt
pymongo==4.6.1
requests==2.31.0
openai>=1.0.0
python-dotenv==1.0.1
httpx>=0.24.0
